wip(scripts): with typescript
