+++ package.json
@@ -3,7 +3,9 @@
   "version": "1.2.0",
   "description": "Js commands for docker stack",
   "scripts": {
-    "help": "node index.js --help"
+    "build": "tsc",
+    "help": "node index.js --help",
+    "publish": "npm run buil && npm publish"
   },
   "files": [
     "phar.json",
@@ -40,5 +42,9 @@
     "readme-md-generator": "1.0.0",
     "semantic-git-commit-cli": "^3.0.0",
     "yaml": "^2.1.3"
+  },
+  "devDependencies": {
+    "@types/node": "^18.11.18",
+    "typescript": "^4.9.4"
   }
 }
